+++ assets/blockly/blockly_editor.css
@@ -1,0 +1,105 @@
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+ul,
+li {
+  list-style: none;
+}
+
+/* 기본 세팅 값 */
+
+.blocklyScrollbarHorizontal,
+.blocklyScrollbarVertical {
+  opacity: 0 !important;
+}
+
+.blocklyScrollbarKnob {
+  opacity: 0 !important;
+}
+
+.scroll-container::-webkit-scrollbar {
+  display: none;
+}
+
+.blocklyZoomOut,
+.blocklyZoomIn {
+  display: none;
+}
+
+.blocklyZoom.blocklyZoomReset {
+  transform: translate(0, 0);
+}
+
+/* 툴박스 영역역 */
+
+.blocklyToolbox {
+  overflow-x: scroll !important;
+  overflow-y: hidden !important;
+  display: flex;
+  align-items: center;
+  background: transparent;
+  white-space: nowrap;
+  -webkit-overflow-scrolling: touch;
+  scroll-behavior: smooth;
+  padding: 8px 0;
+  /* bottom: 133px !important; */
+}
+
+/* 블럭 카테고리 리스트 */
+
+.blocklyToolboxCategoryContainer {
+  margin: 0px !important;
+}
+
+.blocklyToolboxCategoryGroup {
+  flex-wrap: nowrap;
+  gap: 8px;
+  padding: 0px 16px;
+}
+
+.blocklyToolboxCategory {
+  height: 32px !important;
+  display: inline-block;
+  margin-bottom: 0px;
+  padding: 0px 8px !important;
+  background-color: white !important;
+  border-width: 2px !important;
+  border-style: solid !important;
+  border-color: rgba(112, 115, 124, 0.22) !important;
+  border-radius: 16px;
+  box-sizing: border-box;
+}
+
+.blocklyToolboxSelected {
+  background-color: white !important;
+  border-color: #03c75a !important;
+}
+
+.blocklyToolboxCategoryLabel {
+  font-family: "Roboto", sans-serif;
+  font-size: 16px;
+  line-height: 1.2;
+  color: rgba(55, 56, 60, 0.28);
+}
+
+.blocklyToolboxSelected .blocklyToolboxCategoryLabel {
+  color: #03c75a;
+}
+
+/* 블럭 리스트 */
+
+.blocklyFlyout,
+.blocklyToolboxFlyout {
+  height: 133px !important;
+  bottom: 32px;
+  transform: translateY(0) !important;
+}
+
+.blocklyFlyoutBackground {
+  height: 133px !important;
+  background-color: transparent;
+  fill: transparent !important;
+}
