--- assets/blockly/hello.html
@@ -1,140 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-    <head>
-        <meta charset="UTF-8">
-        <script src="https://unpkg.com/blockly/blockly.min.js"></script>
-        <link href="blockly_style.css" rel="stylesheet" />
-    </head>
-    <body>
-        <div id="blocklyDiv" style="height: 100vh; width: 100%;"></div>
-        <xml id="toolbox" style="display: none">
-            <!-- Logic -->
-            <category name="Logic" colour="#5C81A6">
-                <block type="controls_if"></block>
-                <block type="logic_compare"></block>
-                <block type="logic_operation">
-                    <field name="OP">AND</field>
-                </block>
-                <block type="logic_operation">
-                    <field name="OP">OR</field>
-                </block>
-                <block type="logic_boolean">
-                    <field name="BOOL">TRUE</field>
-                </block>
-                <block type="logic_boolean">
-                    <field name="BOOL">FALSE</field>
-                </block>
-                <block type="logic_negate"></block>
-                <block type="logic_null"></block>
-                <block type="logic_ternary"></block>
-            </category>
-
-            <!-- Math -->
-            <category name="Math" colour="#5CA65C">
-                <block type="math_number"></block>
-                <block type="math_arithmetic"></block>
-                <block type="math_single"></block>
-                <block type="math_trig"></block>
-                <block type="math_constant"></block>
-                <block type="math_number_property"></block>
-                <block type="math_round"></block>
-                <block type="math_on_list"></block>
-                <block type="math_modulo"></block>
-                <block type="math_constrain"></block>
-            </category>
-
-            <!-- Text -->
-            <category name="Text" colour="#5CA68D">
-                <block type="text"></block>
-                <block type="text_print">
-                    <value name="TEXT">
-                        <shadow type="text">
-                            <field name="TEXT">Hello, World!</field>
-                        </shadow>
-                    </value>
-                </block>
-                <block type="text_join"></block>
-                <block type="text_length"></block>
-                <block type="text_isEmpty"></block>
-                <block type="text_indexOf"></block>
-                <block type="text_charAt"></block>
-                <block type="text_getSubstring"></block>
-                <block type="text_changeCase"></block>
-                <block type="text_trim"></block>
-            </category>
-
-            <!-- Loops -->
-            <category name="Loops" colour="#5CA699">
-                <block type="controls_repeat_ext"></block>
-                <block type="controls_whileUntil"></block>
-                <block type="controls_for"></block>
-                <block type="controls_forEach"></block>
-                <block type="controls_flow_statements"></block>
-            </category>
-
-            <!-- Lists -->
-            <category name="Lists" colour="#745CA6">
-                <block type="lists_create_empty"></block>
-                <block type="lists_create_with"></block>
-                <block type="lists_repeat"></block>
-                <block type="lists_length"></block>
-                <block type="lists_isEmpty"></block>
-                <block type="lists_indexOf"></block>
-                <block type="lists_getIndex"></block>
-                <block type="lists_setIndex"></block>
-                <block type="lists_getSublist"></block>
-            </category>
-
-            <!-- Variables -->
-            <category
-                    name="Variables"
-                    custom="VARIABLE"
-                    colour="#A6745C"
-            ></category>
-
-            <!-- Functions -->
-            <category
-                    name="Functions"
-                    custom="PROCEDURE"
-                    colour="#A65C81"
-            ></category>
-        </xml>
-
-
-        <script>
-            const workspace = Blockly.inject('blocklyDiv', {
-              toolbox: document.getElementById('toolbox'),
-              media: './media/',
-              trashcan: true,
-              renderer: 'zelos',
-              toolboxPosition: "end",
-              horizontalLayout: true,
-              zoom: { controls: true, wheel: true, startScale: 0.9 }
-            });
-
-            function getXmlJson() {
-                const xmlDom = Blockly.Xml.workspaceToDom(workspace);
-                const xmlText = Blockly.Xml.domToText(xmlDom);
-                return xmlText;
-            }
-
-            function getJavaScriptCode() {
-                return Blockly.JavaScript.workspaceToCode(workspace);
-            }
-
-            function getJsonObject() {
-                return Blockly.serialization.workspaces.save(workspace);
-            }
-
-            window.addEventListener("message", (event) => {
-                if (event.data === "get_all") {
-                    const result = {
-                        json: getJsonObject(),       // XML 형식
-                        javascript: getJavaScriptCode(), // JS Generator 결과
-                    };
-                    FlutterChannel.postMessage(JSON.stringify(result));
-                }
-            });
-        </script>
-    </body>
-</html>
