+++ assets/blockly/toolbox_skin.js
@@ -1,20 +1,85 @@
 (function(){
   const css = `
-  /* 툴박스 컨테이너(좌측 패널 느낌) */
-  .blocklyToolboxDiv {
-    background: #0b1220 !important;
-    color: #fff !important;
-    border-right: 1px solid #1f2937;
-    width: 220px;             /* 고정폭(내용에 따라 넓혀지면 삭제) */
+  /* 툴박스 컨테이너 */
+  .blocklyScrollbarHorizontal,
+  .blocklyScrollbarVertical {
+    opacity: 0 !important;
   }
-  /* 카테고리 라벨(좌측 트리) */
-  .blocklyTreeLabel { color:#e5e7eb !important; font-size: 13px; }
-  .blocklyTreeRow { height: 28px; }
-  .blocklyTreeSelected .blocklyTreeRow { background:#1f2937 !important; }
-  /* 플라이아웃 배경 */
-  .blocklyFlyoutBackground { fill:#0f172a !important; fill-opacity:1 !important; }
-  /* 플라이아웃 라벨(텍스트) */
-  .blocklyFlyoutLabelText { fill:#9ca3af !important; font-size:12px; }
+
+  .blocklyScrollbarKnob {
+    opacity: 0 !important;
+  }
+
+  .scroll-container::-webkit-scrollbar {
+    display: none;
+  }
+
+  .blocklyZoomOut,
+  .blocklyZoomIn {
+    display: none;
+  }
+
+  .blocklyZoom.blocklyZoomReset {
+    transform: translate(0, 0);
+  }
+
+  /* 툴박스 영역역 */
+
+  .blocklyToolbox {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap !important;
+    overflow-x: scroll !important;
+    overflow-y: hidden !important;
+    align-items: center;
+    background: transparent;
+    white-space: nowrap;
+    -webkit-overflow-scrolling: touch;
+    scroll-behavior: smooth;
+    padding: 8px;
+  }
+
+  /* 블럭 카테고리 리스트 */
+
+  .blocklyToolboxCategoryContainer {
+    margin: 0px !important;
+    padding 0 16px;
+  }
+
+  .blocklyToolboxCategoryGroup {
+    flex-wrap: nowrap;
+    gap: 8px;
+  }
+
+  .blocklyToolboxCategory {
+    display: inline-block;
+    margin-bottom: 0px;
+    border-width: 2px !important;
+    border-style: solid !important;
+    border-color: transparent !important;
+    box-sizing: border-box;
+  }
+
+  .blocklyToolboxCategoryLabel {
+    font-family: "Roboto", sans-serif;
+    font-size: 16px;
+    line-height: 1.2;
+  }
+
+  /* 블럭 리스트 */
+
+  .blocklyFlyout,
+  .blocklyToolboxFlyout {
+    height: 133px !important;
+    bottom: 32px;
+    transform: translateY(0) !important;
+  }
+
+  .blocklyFlyoutBackground {
+    height: 133px !important;
+    background-color: transparent;
+    fill: transparent !important;
+  }
   `;
   const style = document.createElement('style');
   style.textContent = css;
