+++ lib/ui/ma_page/splash/ma_splash_page.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:jjava_flutter/ui/ma_page/holder/dashboard/ma_dashboard.dart';
 import 'package:jjava_flutter/ui/ma_page/holder/question/ma_question_page.dart';
 import 'package:jjava_flutter/ui/ma_page/holder/workspace/ma_workspace_page.dart';
 
@@ -61,7 +60,7 @@ class MaSplashPage extends StatelessWidget {
             onPressed: () {
               Navigator.push(
                 context,
-                MaterialPageRoute(builder: (_) => MaDashboard()),
+                MaterialPageRoute(builder: (_) => MaQuestionPage()),
               );
             },
           ),
