+++ pubspec.yaml
@@ -77,8 +77,6 @@ flutter:
     - assets/icons/
     - assets/animation/
     - assets/fonts/
-    - assets/blockly/mobile/ma_blockly_editor.html
-    - assets/blockly/mobile/ma_blockly_style.css
     - assets/blockly/tablet/ta_blockly_editor.html
     - assets/blockly/tablet/ta_blockly_style.css
     - assets/blockly/media/
